refactored Area to us Loop
git-svn-id: svn://localhost/trunk@626 f2a0fb1d-5e3f-4af0-b388-fe59793aa704

+++ src/java/ironfist/Creature.java
@@ -226,7 +226,7 @@ public class Creature implements Serializable {
    */
   public List<Tile> getVision() {
     Vector2D location = getCoordinate();
-    return fov.getSeen(getLevel(), (int) location.getX(),
-      (int) location.getY(), 3);
+    return fov.getSeen(getLevel(), location.getX(),
+      location.getY(), 3);
   }
 }
